fix(project): Fixed the edge detect causing the pixel buffer to become -nan everywhere, just had to swap the two display buffer pointers. In the process, thought switching everything from double to float would maybe help. Hope I didn't screw things up. Contributors: tyler147, hende702

+++ src/highlighter.cc
@@ -41,7 +41,7 @@ void THighlighter::ApplyTool(PixelBuffer* buff, ColorData current_color,
     int mid_y = height / 2;
     int screen_h = buff->height();
     int screen_w = buff->width();
-    double** cur_mask = this->get_mask();
+    float** cur_mask = this->get_mask();
 
     /* Check the current mask appliation with previous
     mask application to avoid intense opacity */
